refactor dir structure
